added a actual or future bookings check, to avoid that user can delete listing with future or actual bookings with status CONFIRMED

+++ src/main/java/com/java2024/ecoscape/services/ListingService.java
@@ -5,15 +5,14 @@ import com.java2024.ecoscape.dto.ListingResponse;
 import com.java2024.ecoscape.dto.ListingRulesDTO;
 import com.java2024.ecoscape.models.Listing;
 import com.java2024.ecoscape.models.Rules;
+import com.java2024.ecoscape.models.Status;
 import com.java2024.ecoscape.models.User;
-import com.java2024.ecoscape.repositories.ListingAvailableDatesRepository;
-import com.java2024.ecoscape.repositories.ListingRepository;
-import com.java2024.ecoscape.repositories.RulesRepository;
-import com.java2024.ecoscape.repositories.UserRepository;
+import com.java2024.ecoscape.repositories.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -27,14 +26,16 @@ public class ListingService {
     private final RulesService rulesService;
     private final ListingAvailableDatesService listingAvailableDatesService;
     private final ListingAvailableDatesRepository listingAvailableDatesRepository;
+    private final BookingRepository bookingRepository;
 
-    public ListingService(ListingRepository listingRepository, UserRepository userRepository, RulesRepository rulesRepository, RulesService rulesService, ListingAvailableDatesService listingAvailableDatesService, ListingAvailableDatesRepository listingAvailableDatesRepository) {
+    public ListingService(ListingRepository listingRepository, UserRepository userRepository, RulesRepository rulesRepository, RulesService rulesService, ListingAvailableDatesService listingAvailableDatesService, ListingAvailableDatesRepository listingAvailableDatesRepository, BookingRepository bookingRepository) {
         this.listingRepository = listingRepository;
         this.rulesRepository = rulesRepository;
         this.userRepository = userRepository;
         this.rulesService = rulesService;
         this.listingAvailableDatesService = listingAvailableDatesService;
         this.listingAvailableDatesRepository = listingAvailableDatesRepository;
+        this.bookingRepository = bookingRepository;
     }
 
     //metod för att skaffa listing
@@ -125,8 +126,12 @@ public class ListingService {
         return listingRepository.save(listing);
     }
     //metod for att ta bort en listing
+    @Transactional
     public void deleteListingById(Long id){
         Listing listing = listingRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Listig not found"));
+        if(bookingRepository.existsByListingIdAndEndDateAfterAndStatus(id, LocalDate.now(), Status.CONFIRMED)){
+            throw new IllegalArgumentException("You cannot delete the listing with actual or future booking!");
+        }
         //om listingen har availabible dates, ska de tas bort innan listingen tas bort
         if (listingAvailableDatesRepository.existsByListingId(id)) {
             listingAvailableDatesService.deleteAllAvailableDatesByHostOfAListing(id);
